added deletion functionality on the server for chats

+++ server/Controller/chatController.js
@@ -1,38 +1,68 @@
 const chatModel = require("../Models/chatModel");
+const messageModel = require("../Models/messageModel");
 
 const createChat = async (req, res) => {
-  const { firstId, secondId } = req.body;
+  const { userId, recipientId } = req.body;
   try {
-    const chat = await chatModel.findOne({
-      members: { $all: [firstId, secondId] },
+    const existingChat = await chatModel.findOne({
+      members: { $all: [userId, recipientId] },
     });
-    if (chat) return res.status(200).json(chat);
+    if (existingChat) {
+      const isUserIdPresent = existingChat.deleteId.includes(userId);
+      if (!isUserIdPresent) {
+        existingChat.deleteId.push(userId);
+        await existingChat.save();
+      }
+      return res.status(200).json(existingChat);
+    }
 
-    const newChat = new chatModel({
-      members: [firstId, secondId],
+    const newChatForUser = new chatModel({
+      members: [userId, recipientId],
+      deleteId: [userId, recipientId],
     });
-    const response = await newChat.save();
+
+    const response = await newChatForUser.save();
+
     res.status(200).json(response);
   } catch (error) {
     res.status(500).json(error);
   }
 };
+
 const getAllUserChats = async (req, res) => {
   const userId = req.params.userId;
+
   try {
-    const userChats = await chatModel.find({
-      members: { $in: [userId] },
-    });
-    res.status(200).json(userChats);
+    const userChats = await chatModel
+      .find({
+        members: userId,
+      })
+      .where("members")
+      .equals(userId);
+    if (userChats) {
+      const isNotDeletedChats = userChats.filter((chat) =>
+        chat.deleteId.includes(userId)
+      );
+      // const uniqueMembersSet = new Set();
+      // const uniqueChat = isNotDeletedChats.filter((obj) => {
+      //   const membersString = JSON.stringify(obj.members);
+      //   if (!uniqueMembersSet.has(membersString)) {
+      //     uniqueMembersSet.add(membersString);
+      //     return true;
+      //   }
+      //   return false;
+      // });
+      res.status(200).json(isNotDeletedChats);
+    }
   } catch (error) {
     res.status(500).json(error);
   }
 };
 const getSingleChat = async (req, res) => {
-  const { firstId, secondId } = req.params;
+  const { userId, recipientId } = req.params;
   try {
     const chat = await chatModel.findOne({
-      members: { $all: [firstId, secondId] },
+      members: [userId, recipientId],
     });
     res.status(200).json(chat);
   } catch (error) {
@@ -40,13 +70,32 @@ const getSingleChat = async (req, res) => {
   }
 };
 const deleteSingleChat = async (req, res) => {
-  const { chatId } = req.params;
+  const { userId, chatId } = req.params;
+
   try {
     const chat = await chatModel.findById(chatId);
 
     if (chat) {
-      const response = await chat.deleteOne();
-      res.status(200).json(response);
+      const remainingId = chat.deleteId.filter((id) => id !== userId);
+      chat.deleteId = remainingId;
+
+      const messages = await messageModel.find({ chatId });
+
+      messages.map(async (message) => {
+        const remainingId = message.deleteId.filter((id) => id !== userId);
+        message.deleteId = remainingId;
+
+        await message.save();
+      });
+
+      if (remainingId.length === 0) {
+        await chat.deleteOne();
+        await messageModel.deleteMany({ chatId });
+        res.status(200).json({ data: "Chat deleted", status: 200 });
+      } else {
+        const response = await chat.save();
+        res.status(200).json({ data: response, status: 200 });
+      }
     } else {
       res.status(404).json({ message: "Chat not found" });
     }
